LayoutTests/http/tests/websocket/tests/hixie76/split-binary-frame-header-expected.txt

WebSocket: Hixie76 binary frames should be handled correctly even if frame header is split into multiple TCP packets.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Connected.
Error occurred.
Closed.
PASS closeEvent.wasClean is true
PASS errorCount is 1
PASS successfullyParsed is true

TEST COMPLETE
